+++ packages/bits-ui/src/lib/bits/select/select.svelte.ts
@@ -18,6 +18,7 @@ import {
 	boolToEmptyStrOrUndef,
 	getDataOpenClosed,
 	boolToTrueOrUndef,
+	getDataTransitionAttrs,
 } from "$lib/internal/attrs.js";
 import { kbd } from "$lib/internal/kbd.js";
 import type {
@@ -112,6 +113,7 @@ abstract class SelectBaseRootState {
 		if (!this.highlightedNode) return null;
 		return this.highlightedNode.getAttribute("data-label");
 	});
+	contentIsPositioned = $state(false);
 	isUsingKeyboard = false;
 	isCombobox = false;
 	domContext = new DOMContext(() => null);
@@ -138,8 +140,13 @@ abstract class SelectBaseRootState {
 	setHighlightedNode(node: HTMLElement | null, initial = false) {
 		this.highlightedNode = node;
 		if (node && (this.isUsingKeyboard || initial)) {
-			node.scrollIntoView({ block: this.opts.scrollAlignment.current });
+			this.scrollHighlightedNodeIntoView(node);
 		}
+	}
+
+	scrollHighlightedNodeIntoView(node: HTMLElement) {
+		if (!this.viewportNode || !this.contentIsPositioned) return;
+		node.scrollIntoView({ block: this.opts.scrollAlignment.current });
 	}
 
 	getCandidateNodes(): HTMLElement[] {
@@ -903,6 +910,7 @@ export class SelectContentState {
 
 		onDestroyEffect(() => {
 			this.root.contentNode = null;
+			this.root.contentIsPositioned = false;
 			this.isPositioned = false;
 		});
 
@@ -910,9 +918,15 @@ export class SelectContentState {
 			() => this.root.opts.open.current,
 			() => {
 				if (this.root.opts.open.current) return;
+				this.root.contentIsPositioned = false;
 				this.isPositioned = false;
 			}
 		);
+
+		watch([() => this.isPositioned, () => this.root.highlightedNode], () => {
+			if (!this.isPositioned || !this.root.highlightedNode) return;
+			this.root.scrollHighlightedNodeIntoView(this.root.highlightedNode);
+		});
 
 		this.onpointermove = this.onpointermove.bind(this);
 	}
@@ -962,6 +976,7 @@ export class SelectContentState {
 				role: "listbox",
 				"aria-multiselectable": this.root.isMulti ? "true" : undefined,
 				"data-state": getDataOpenClosed(this.root.opts.open.current),
+				...getDataTransitionAttrs(this.root.contentPresence.transitionStatus),
 				[this.root.getBitsAttr("content")]: "",
 				style: {
 					display: "flex",
@@ -987,6 +1002,7 @@ export class SelectContentState {
 			// onPlaced is also called when the menu is closed, so we need to check if the menu
 			// is actually open to avoid setting positioning to true when the menu is closed
 			if (this.root.opts.open.current) {
+				this.root.contentIsPositioned = true;
 				this.isPositioned = true;
 			}
 		},
@@ -1427,7 +1443,8 @@ export class SelectScrollDownButtonState {
 				}
 				this.scrollIntoViewTimer = afterSleep(5, () => {
 					const activeItem = this.root.highlightedNode;
-					activeItem?.scrollIntoView({ block: this.root.opts.scrollAlignment.current });
+					if (!activeItem) return;
+					this.root.scrollHighlightedNodeIntoView(activeItem);
 				});
 			}
 		);