+++ packages/bits-ui/CHANGELOG.md
@@ -1,5 +1,33 @@
 # bits-ui
 
+## 2.16.5
+
+### Patch Changes
+
+- fix(DateField): `onbeforeinput` handling ([#1990](https://github.com/huntabyte/bits-ui/pull/1990))
+
+- fix(Select): scroll jumping ([#2005](https://github.com/huntabyte/bits-ui/pull/2005))
+
+- fix(Tooltip): skip delay race condition with multiple tooltips ([#2002](https://github.com/huntabyte/bits-ui/pull/2002))
+
+- fix(NavigationMenu): delay hover close; add leave-delay regression test ([#1983](https://github.com/huntabyte/bits-ui/pull/1983))
+
+## 2.16.4
+
+### Patch Changes
+
+- fix: menu not ignoring onCloseAutoFocus when closed as another focus scope ([#1999](https://github.com/huntabyte/bits-ui/pull/1999))
+
+- fix: floating layer bouncing on scroll ([#1997](https://github.com/huntabyte/bits-ui/pull/1997))
+
+## 2.16.3
+
+### Patch Changes
+
+- fix: floating content exit animations ([#1980](https://github.com/huntabyte/bits-ui/pull/1980))
+
+- fix(Tooltip): improve transit handling from trigger to content ([#1981](https://github.com/huntabyte/bits-ui/pull/1981))
+
 ## 2.16.2
 
 ### Patch Changes